Attempt to init the underglow device with the PM define
This commit doesn’t work, because the linker complains about multiple defines of the object.

That’s because `zephyr/drivers/led_strip/ws2812_spi.c` also calls a similar DEFINE function and we end up with a duplicate.

But that function doesn’t do anything with power domains.

Perhaps that’s why the power domain is not added?